requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
websockets==11.0.3
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
xrpl-py==3.0.0
python-socketio==5.4.0
aiohttp==3.8.1
httpx==0.24.1
ccxt==4.4.35
